+++ pkg/util/fputil/dirsize.go
@@ -1,0 +1,17 @@
+package fputil
+
+import (
+	"io/fs"
+	"path/filepath"
+)
+
+func DirectorySize(path string) int64 {
+	var size int64
+	filepath.Walk(path, func(_ string, info fs.FileInfo, _ error) error {
+		if info.Mode().IsRegular() {
+			size += info.Size()
+		}
+		return nil
+	})
+	return size
+}